# api/api_server/swagger.yaml
definitions:
...
  api.ApiPaginatorReviews:
    properties:
      count:
        type: integer
      page_num:
        type: integer
      sort_direction:
        type: string
      sort_key:
        type: string
    type: object
...
  api.ApiRangeReviews:
    properties:
      list_reviews:
        items:
          $ref: '#/definitions/usecase.ViewReview'
        type: array
      max_count_pages:
        type: integer
    type: object
  api.ApiReview:
    properties:
      advantages:
        type: string
      comment:
        type: string
      disadvantages:
        type: string
      is_public:
        type: boolean
      product_id:
        type: integer
      rating:
        type: integer
    type: object
  api.ApiReviewStatistics:
    properties:
      stars:
        items:
          type: integer
        type: array
    type: object
...
  errors.Error:
    properties:
      error:
        type: string
    type: object
...
  usecase.ViewReview:
    properties:
      advantages:
        type: string
      comment:
        type: string
      date_added:
        type: string
      disadvantages:
        type: string
      rating:
        type: integer
      user_avatar:
        type: string
      user_name:
        type: string
    type: object
info:
  contact: {}
paths:
...
  /review/product:
    post:
      consumes:
      - application/json
      description: Добавление нового отзыва.
      parameters:
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      - description: Данные отзыва.
        in: body
        name: Review
        required: true
        schema:
          $ref: '#/definitions/api.ApiReview'
      produces:
      - application/json
      responses:
        "200":
          description: Отзыв успешно добавлен.
          schema:
            $ref: '#/definitions/errors.Error'
        "400":
          description: Некорректное тело запроса.
          schema:
            $ref: '#/definitions/errors.Error'
        "500":
          description: Непредвиденная ошибка сервера.
          schema:
            $ref: '#/definitions/errors.Error'
      security:
      - ApiKeyAuth: []
      summary: Добавление нового отзыва.
  /review/product/{id}:
    post:
      consumes:
      - application/json
      description: Получение отзывов для определённого товара.
      parameters:
      - description: Данные товара.
        in: body
        name: PaginatorReviews
        required: true
        schema:
          $ref: '#/definitions/api.ApiPaginatorReviews'
      produces:
      - application/json
      responses:
        "200":
          description: Отзывы успешно получены.
          schema:
            $ref: '#/definitions/api.ApiRangeReviews'
        "400":
          description: Некорректное тело запроса.
          schema:
            $ref: '#/definitions/errors.Error'
        "500":
          description: Непредвиденная ошибка сервера.
          schema:
            $ref: '#/definitions/errors.Error'
      summary: Получение отзывов для определённого товара.
  /review/rights/product/{id}:
    get:
      consumes:
      - application/json
      description: Проверка прав для написания отзыва.
      parameters:
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      - description: Id товара
        in: query
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Проверка прав для написания отзыва.
          schema:
            $ref: '#/definitions/api.ApiReviewStatistics'
        "400":
          description: Некорректное тело запроса.
          schema:
            $ref: '#/definitions/errors.Error'
        "500":
          description: Непредвиденная ошибка сервера.
          schema:
            $ref: '#/definitions/errors.Error'
      security:
      - ApiKeyAuth: []
      summary: Проверка прав для написания отзыва.
  /review/statistics/product/{id}:
    post:
      consumes:
      - application/json
      description: Получение подробного рейтинга товара.
      parameters:
      - description: Id товара
        in: query
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Получение подробного рейтинга.
          schema:
            $ref: '#/definitions/api.ApiReviewStatistics'
        "400":
          description: Некорректное тело запроса.
          schema:
            $ref: '#/definitions/errors.Error'
        "500":
          description: Непредвиденная ошибка сервера.
          schema:
            $ref: '#/definitions/errors.Error'
      summary: Получение подробного рейтинга товара.
...
swagger: "2.0"